Struct epub_builder::TocElement

pub struct TocElement {
    pub level: i32,
    pub url: String,
    pub title: String,
    pub children: Vec<TocElement>,
}

An element of the Table of contents

Example

use epub_builder::TocElement;
TocElement::new("chapter_1.xhtml", "Chapter 1")
    .child(TocElement::new("chapter_1.xhtml#1", "Chapter 1, section 1")
              .child(TocElement::new("chapter_1.xhtml#1-1", "Chapter 1, section 1, subsection 1")));

Fields
level: i32

The level. 0: part, 1: chapter, 2: section, ...

url: String

The link

title: String

Title of this entry

children: Vec<TocElement>

Inner elements

Methods

impl TocElement

pub fn new<S1: Into<String>, S2: Into<String>>(url: S1, title: S2) -> TocElement

Creates a new element of the toc

By default, the element's level is 1 and it has no children.

pub fn level(self, level: i32) -> Self

Sets the level of a TocElement

pub fn child(self, child: TocElement) -> Self

Add a child to this element.

This adjust the level of the child to be the level of its parents, plus 1; this means that there is no point in manually setting the level to elements added with this method.

Example

use epub_builder::TocElement;
let elem = TocElement::new("foo.xhtml", "Foo")
    .child(TocElement::new("bar.xhtml", "Bar")
         .level(42));

// `Bar`'s level wiss still be `2`.

pub fn add(&mut self, element: TocElement)

Add element to self or to children, according to its level

This will adds element directly to self if its level is equal or less to the last children element; else it will insert it to the last child.

See the add method of `Toc.