Struct epub_builder::ZipCommand

pub struct ZipCommand { /* fields omitted */ }

Zip files using the system zip command.

Create a temporary directory, write temp files in that directory, and then calls the zip command to generate an epub file.

This method will fail if zip (or the alternate specified command) is not installed on the user system.

Note that these takes care of adding the mimetype (since it must not be deflated), it should not be added manually.

Methods

impl ZipCommand

pub fn new() -> Result<ZipCommand>

Creates a new ZipCommand, using default setting to create a temporary directory.

pub fn new_in<P: AsRef<Path>>(temp_path: P) -> Result<ZipCommand>

Creates a new ZipCommand, specifying where to create a temporary directory.

Arguments

temp_path: the path where a temporary directory should be created.

pub fn command<S: Into<String>>(&mut self, command: S) -> &mut Self

Set zip command to use (default: "zip")